chore: use interface extends instead of declare module
